[refactor](be) Remove redundant casts from BE code (#64087)
### What problem does this PR solve?


Problem Summary: Remove redundant casts reported by clang-tidy
readability-redundant-casting in BE code. These casts were no-op
conversions to the same type, including redundant static_cast,
reinterpret_cast, and C-style casts. The change keeps casts that still
perform real type conversion and only removes the redundant outer casts
reported by clang-tidy.

### Release note

None

### Release note

None

+++ be/src/format/orc/vorc_reader.cpp
@@ -748,7 +748,7 @@ std::tuple<bool, orc::Literal> convert_to_orc_literal(const orc::Type* type,
             } else if constexpr (primitive_type == TYPE_INT) {
                 return std::make_tuple(true, orc::Literal(int64_t(*((int32_t*)value))));
             } else if constexpr (primitive_type == TYPE_BIGINT) {
-                return std::make_tuple(true, orc::Literal(int64_t(*((int64_t*)value))));
+                return std::make_tuple(true, orc::Literal(*((int64_t*)value)));
             }
             return std::make_tuple(false, orc::Literal(false));
         }
@@ -756,7 +756,7 @@ std::tuple<bool, orc::Literal> convert_to_orc_literal(const orc::Type* type,
             if constexpr (primitive_type == TYPE_FLOAT) {
                 return std::make_tuple(true, orc::Literal(double(*((float*)value))));
             } else if constexpr (primitive_type == TYPE_DOUBLE) {
-                return std::make_tuple(true, orc::Literal(double(*((double*)value))));
+                return std::make_tuple(true, orc::Literal(*((double*)value)));
             }
             return std::make_tuple(false, orc::Literal(false));
         }
